move thinning params from thunk to dataview config
